% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ASDEPslbenef}
\alias{ASDEPslbenef}
\title{Bénéficiaires d'aides sociales départementales}
\format{
Un data frame avec 3125 lignes et 51 variables:
\describe{
\item{Annee}{année (format character)}
\item{Code.region}{code numérique de la région}
\item{Code.departement}{code numérique du département (ou du territoire gérant l'aide sociale)}
\item{TypeTerritoire}{type de territoire ("département","région",...)}
\item{Territoire}{nom complet du territoire}
\item{nom.variable}{voir la table ASDEPslbenef_description}
}
}
\source{
\url{https://data.drees.solidarites-sante.gouv.fr/explore/dataset/375_les-beneficiaires-de-l-aide-sociale-departementale/information/}
}
\usage{
ASDEPslbenef
}
\description{
Une base de données issues de l'enquête annuelle 'Aide sociale' de la DREES, contenant les effectifs de bénéficiaires de diverses aides sociales versées par les départements.
Les noms des variables, correspondant aux différents types d'aide, sont décrits dans la table 'ASDEPslbenef_description'.
}
\keyword{datasets}
