% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ASDEPsldepenses}
\alias{ASDEPsldepenses}
\title{Dépenses annuelles d'aides sociales départementales}
\format{
Un data frame avec 2460 lignes et 38 variables:
\describe{
\item{Annee}{année (format character)}
\item{Code.region}{code numérique de la région}
\item{Code.departement}{code numérique du département (ou du territoire gérant l'aide sociale)}
\item{TypeTerritoire}{type de territoire ("département","région",...)}
\item{Territoire}{nom complet du territoire}
\item{nom.variable}{voir la table ASDEPsldepenses_description}
}
}
\source{
\url{http://www.data.drees.sante.gouv.fr/}
}
\usage{
ASDEPsldepenses
}
\description{
Une base de données issues de l'enquête annuelle 'Aide sociale' de la DREES, contenant les montants annuels de dépenses de diverses aides sociales versées par les départements.
Les noms des variables, correspondant aux différents types d'aide, sont décrits dans la table 'ASDEPsldepenses_description'.
}
\keyword{datasets}
