% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ASDEPsldepenses_description}
\alias{ASDEPsldepenses_description}
\title{Intitulés des variables des bases sur les dépenses d'aides sociales départementales}
\format{
Un data frame avec 34 lignes et 13 variables:
\describe{
\item{nom.var}{nom de la variable (correspond aux colonnes des bases ASDEPslbenef et ASDEPsldepenses)}
\item{Intitule.var}{intitulé complet de la variable}
\item{Intitulecourt.var}{intitulé court de la variable}
\item{Source.var}{source (= enquête Aide sociale, DREES)}
\item{Champ.var}{champ de la variable}
\item{Note.var}{notes explicatives sur la variable, récupérées du fichier Excel diffusé par la DREES}
\item{Unite.var}{unité de la variable (personnes, €, ...)}
\item{Thematique.var}{thématique (perte d'autonomie, handicap, ...)}
\item{TexteDenom.var}{texte court lorsque la variable est utilisé comme dénominateur d'un ratio}
\item{ListeDenom.var}{liste des noms d'autres variables (séparés par un '\emph{') pouvant servir comme dénominateur de la variable}
\item{ListeComposante.var}{listes de noms d'autres variables (séparés par un '}') dont la somme correspond à la variable}
\item{Type.var}{type de variable (nombre, dépense...)}
\item{Popref.var}{population de référence (par exemple : "20-59" = les 20-59 ans)}
}
}
\source{
\url{http://www.data.drees.sante.gouv.fr/}
}
\usage{
ASDEPsldepenses_description
}
\description{
Une base de métadonnées sur les variables de la base ASDEPsldepenses
}
\keyword{datasets}
