% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/champFrance.R
\name{champFrance}
\alias{champFrance}
\title{Une fonction auxiliaire qui retourne le champ le plus large possible pour lequel une variable est non nulle}
\usage{
champFrance(var, annee = max(ASDEPsl$Annee), tab = ASDEPsl)
}
\arguments{
\item{var}{un nom de variable de la table ASDEPsl}

\item{annee}{une année}
}
\value{
un nom de territoire
}
\description{
Si la variable est disponible pour le champ France entière, ce champ est retenu, puis dans l'ordre : France entière hors Mayotte, France métropolitaine.
}
\examples{
champFrance("NbBenefAPA")
champFrance("NbBenefAPA",2005)
}
