% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{departementsFR}
\alias{departementsFR}
\title{Liste des départements et territoires}
\format{
Un data frame avec 104 lignes et 4 variables
\describe{
\item{NumReg}{Code numérique de la région}
\item{NumDept}{Code alphnumérique du département (ou territoire)}
\item{Departement}{Nom du département}
\item{Region}{Nom de la région}
}
}
\source{
\url{}
}
\usage{
departementsFR
}
\description{
Une base de données contenant la liste des départements (et territoires gérant l'aide sociale départementale)
}
\keyword{datasets}
