% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/departementsProches.R
\name{departementsProches}
\alias{departementsProches}
\title{Une fonction auxiliaire déterminant les départements les plus proches d'un département de référence au regard d'un indicateur donné}
\usage{
departementsProches(
  dept = NULL,
  nomvariable = NULL,
  denom = "",
  nb = 5,
  annee = max(ASDEPsl[is.na(ASDEPsl[, c(nomvariable1, nomvariable2)]), "Annee"], na.rm
    = TRUE),
  tabindic = NULL,
  donnees = ASDEPsl,
  variables = ASDEPsl_description
)
}
\arguments{
\item{dept}{le nom du département de référence}

\item{nomvariable}{le nom de la variable de référence (sur la bsse de laquelle les départements les plus proches seront déterminés)}

\item{denom}{l'éventuel dénominateur de la variable}

\item{nb}{le nombre de départements proches à inclure dans le vecteur en sortie}

\item{annee}{l'année de référence (par défaut, la dernière année pour laquelle la variable est disponible)}

\item{tabindic}{une table avec les valeurs départementales de l'indicateur (si cette table n'est pas fournie en entrée, elle est construire par la fonction)}

\item{donnees}{les données de base (par défaut, la table ASDEPsl contenue dans le package asdep)}

\item{variables}{les métadonnées de base (par défaut, la table ASDEPsl_description)}
}
\value{
un vecteur de noms de départements
}
\description{
Fonction qui retourne un vecteur des départements les plus proches (variable 'nb', par défaut 5 départements) du département de référence (variable 'dept') au regard d'un indicateur donné (variable 'nomvariable')
}
\examples{
departementsProches(dept="Meurthe-et-Moselle",nomvariable="NbBenefAPA",denom="popPA",nb=5)
}
