% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotAsdep.R
\name{ggplotAsdep}
\alias{ggplotAsdep}
\title{Une fonction auxiliaire permettant de mettre en forme des graphiques ggplot}
\usage{
ggplotAsdep(...)
}
\arguments{
\item{...}{les paramètres en input de la fonction ggplot}
}
\value{
un graphique ggplot
}
\description{
Une fonction auxiliaire permettant de mettre en forme des graphiques ggplot
}
