% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotlyAsdep.R
\name{ggplotlyAsdep}
\alias{ggplotlyAsdep}
\title{Une fonction auxiliaire permettant de mettre en forme des graphiques dynamique plotly}
\usage{
ggplotlyAsdep(...)
}
\arguments{
\item{...}{les paramètres en input de la fonction ggplot}
}
\value{
un graphique plotly
}
\description{
Une fonction auxiliaire permettant de mettre en forme des graphiques dynamique plotly
}
