% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphComparaison.R
\name{graphComparaison}
\alias{graphComparaison}
\title{Fonction produisant un graphique comparant, pour une année donnée, les valeurs par département d'un indicateur d'aide sociale}
\usage{
graphComparaison(
  nomvariable,
  denom = "",
  annee = max(ASDEPsl[!is.na(ASDEPsl[, nomvariable]), "Annee"], na.rm = TRUE),
  options = c(),
  poidsobs = c(),
  dept,
  comp = "TOTAL estimé France entière (hors Mayotte)",
  gpecomp = c(),
  typesortie = "graph",
  donnees = ASDEPsl,
  variables = ASDEPsl_description
)
}
\arguments{
\item{nomvariable}{le nom de la variable, parmi celles de la table ASDEPsl}

\item{denom}{le nom d'une variable à utiliser au dénominateur de l'indicateur (si absent la variable est utilisée brute)}

\item{annee}{l'année de référence pour le calcul de l'indicateur (par défaut, l'année la plus récente pour laquelle l'indicateur 'nomvariable' est connu)}

\item{options}{un vecteur d'options du calcul (valeurs en euros courant ou constant, mensuels ou annuels, etc.) et d'options de représentation (représenter les zones interdécile, interquartile, médiane +/- 10 \%, etc)}

\item{poidsobs}{une variable de pondération (par exemple si on souhaite calculer des quantiles pondérés par la population des départements). Par défaut, si denom est renseigné, cette variable denom est utilisée comme poidsobs}

\item{dept}{le nom du département qu'on souhaite représenter}

\item{comp}{le nom du territoire qu'on souhaite représenter en comparaison (par défaut, "TOTAL estimé France entière (hors Mayotte)")}

\item{gpecomp}{un vecteur de noms de départements : leur ensemble constituera le "groupe de comparaison"}

\item{typesortie}{détermine l'objet en sortie de la fonction : un graphique si l'option "graph" est retenue (option par défaut), une table de donnée (data frame) si l'option "tab" est retenue. Les options "tabcomplet" (une liste de tables) et "graphdyn" (un graphique Plotly) sont également disponibles}

\item{donnees}{la table de données en entrée (par défaut, la table ASDEPsl)}

\item{variables}{la table de métadonnées de la table de données en entrée (par défaut, la table ASDEPsl_description)}
}
\value{
un graphique (format ggplot ou plotly), ou éventuellement un tableau (selon les options retenues)
}
\description{
Cette fonction sert à produire un graphique (au format ggplot2 ou plotly) représentant l'évolution au cours du temps d'un indicateur pour un département de référence, et un territoire de comparaisons (par exemple : la France entière, la région du département de référence, un groupe de département).
Le tableau sous-jacent peut également être récupéré avec l'option typesortie = "tab".
Les options permettent d'ajouter sur le graphique une représentation des diverses zones : zone interquartile ("interquartiles"), zone interdécile ("interdeciles"), zone correspondant à la médiane +/- 10 ou 20 \% ("medianePM10" et "medianePM20") ...
}
\examples{
graphComparaison(nomvariable="NbBenefAPA",dept="Vosges",comp="Grand Est",typesortie="tab")
graphComparaison(nomvariable="NbBenefAPA",denom="pop.60.99",dept="Vosges",comp="Grand Est",options=c("interquartiles","interdeciles","medianePM10"))
graphComparaison(nomvariable="NbBenefAPA",denom="pop.60.99",dept="Vosges",gpecomp=c("Meuse","Moselle"),options=c("medianePM10","medianePM20"),typesortie="graphdyn")
}
