% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intitulepop.R
\name{Intitulepop}
\alias{Intitulepop}
\title{Fonction auxiliaire associant un intitulé en clair à un nom de variable de population}
\usage{
Intitulepop(nom)
}
\arguments{
\item{nom}{un nom de variable de population (dans la table PopDepartementales)}
}
\value{
un intitulé en clair
}
\description{
Fonction auxiliaire associant un intitulé en clair à un nom de variable de population
}
\examples{
Intitulepop("pop.20.59")
Intitulepop("pop.60.99")
Intitulepop("popTOT")
Intitulepop("popASE")
}
