% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LitOngletOarsa.R
\name{LitOngletOarsa}
\alias{LitOngletOarsa}
\title{Fonction auxiliaire extrayant les données d'un onglet du fichier Excel de résultats de l'enquête OARSA (DREES)}
\usage{
LitOngletOarsa(
  Nom.var,
  nomfich = FichierSource,
  nomsheet,
  corrigenom = function(x) {     x }
)
}
\arguments{
\item{Nom.var}{le nom de la variable en sortie, dans la table de donnée produite}

\item{nomfich}{le nom du fichier Excel dans lequel se trouvent les données}

\item{nomsheet}{le nom de l'onglet du fichier Excel}

\item{corrigenom}{une fonction corrigeant les noms de département}
}
\value{
une liste de deux tables : "tab" = la table des données, "infovar" = la table des métadonnées (intitulé complet, source, unité, etc.)
}
\description{
Cette fonction auxiliaire sert à la création d'une table de données (format Rds) à partir des données
mises en ligne au format Excel sur data.drees pour la diffusion des résultats de l'enquête annuelle OARSA
(orientation et accompagnement des bénéficiaires du RSA).
Elle est adaptée au format du fichier disponible en novembre 2020.
}
