% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OARSAsl}
\alias{OARSAsl}
\title{Orientation et accompagnement des bénéficiaires du RSA}
\format{
Un data frame avec 416 lignes et 23 variables:
\describe{
\item{Annee}{année (format character)}
\item{TypeTerritoire}{type de territoire ("département","région",...)}
\item{Territoire}{nom complet du territoire}
\item{nom.variable}{voir la table OARSAsl_description}
}
}
\source{
\url{http://www.data.drees.sante.gouv.fr/}
}
\usage{
OARSAsl
}
\description{
Une base de données issues de l'enquête annuelle 'Orientation et accompagnement des bénéficiaires du RSA' de la DREES,
contenant des indicateurs sur les parts de bénéficiaires orientés par les départements et les délais moyens d'orientation.
Les noms des variables, correspondant aux différents types d'aide, sont décrits dans la table 'OARSAsl_description'.
}
\keyword{datasets}
