% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optionsgraphiques.R
\name{optionsgraphiques}
\alias{optionsgraphiques}
\title{Une fonction auxiliaires qui fournit les options d'affichage d'un graphique départemental}
\usage{
optionsgraphiques(
  dept,
  comp = "TOTAL estimé France entière (hors Mayotte)",
  options = c()
)
}
\arguments{
\item{dept}{le nom du département qu'on souhaite représenter}

\item{comp}{le nom du territoire qu'on souhaite représenter en comparaison (par défaut, "TOTAL estimé France entière (hors Mayotte)")}

\item{options}{un vecteur d'options issu de la fonction d'origine}
}
\value{

}
\description{
Cette fonction est notamment utilisée dans les fonctions graphEvolution, graphComparaison, etc.
}
