% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PopDepartementales}
\alias{PopDepartementales}
\title{Populations départementales}
\format{
Un data frame avec 2646 lignes et 12 variables
}
\source{
\url{https://www.insee.fr/fr/statistiques/1893198}
}
\usage{
PopDepartementales
}
\description{
Une base de données contenant les populations départementales par tranche d'âge, téléchargéex sur le site de l'Insee.
}
\keyword{datasets}
