% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantileIndic.R
\name{quantileIndic}
\alias{quantileIndic}
\title{Une fonction auxiliaire calculant les quantiles des distributions par département d'un indicateur}
\usage{
quantileIndic(
  donneesQV,
  var,
  groupe,
  poids,
  liste.quantiles = c(0.05, 0.1, 0.25, 0.5, 0.75, 0.9, 0.95)
)
}
\arguments{
\item{donneesQV}{une table de données contant la variable d'intérêt (var),}

\item{var}{le nom de variable pour l'indicateur dont on souhaite calculer les quantiles}

\item{groupe}{un nom de variable pour segmenter le calcul par groupe (par exemple : "Annee")}

\item{poids}{une variable de pondération pour calculer des quantiles pondérés (par exemple la population : "popTOT")}

\item{liste.quantiles}{un vecteur contenant les quantiles qu'on veut inclure dans les résultats}
}
\value{
une table contenant les quantiles, les nombres de départements (et éventuellement part de la population) dans les zones interquartile et interdécile, ainsi que dans les zones autour de la médiane +/- 10 ou 20 \%, et la variable "groupe"
}
\description{
Une fonction auxiliaire calculant les quantiles des distributions par département d'un indicateur
}
\examples{
quantileIndic( donneesQV=selectIndic(nomvariable="NbBenefAPA",denom="pop.60.99", keepvar = c("pop.60.99"))$var, var="NbBenefAPA", groupe="Annee", poids="pop.60.99")
}
