% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectIndic.R
\name{selectIndic}
\alias{selectIndic}
\title{Une fonction pour extraire et calculer un indicateur sur l'aide sociale}
\usage{
selectIndic(
  nomvariable,
  denom = "",
  options = "",
  gpeDpt = c(),
  donnees = ASDEPsl,
  variables = ASDEPsl_description
)
}
\arguments{
\item{nomvariable}{le nom de la variable, parmi celles de la table ASDEPsl}

\item{denom}{le nom d'une variable à utiliser au dénominateur de l'indicateur (si absent la variable est utilisée brute)}

\item{options}{un vecteur d'options du calcul}

\item{gpeDpt}{un vecteur de noms de départements : leur ensemble constituera le "groupe de comparaison"}

\item{donnees}{la table de données en entrée (par défaut, la table ASDEPsl)}

\item{variables}{la table de métadonnées de la table de données en entrée (par défaut, la table ASDEPsl_description)}
}
\value{
une liste contenant 4 éléments : 'var' = une table avec les territoires, les années et les valeurs de l'indicateur calculé; 'unitevar', 'unitevarbase' et 'nivarrond' = des métadonnées sur l'indicateur calculé
}
\description{
Une fonction pour extraire et calculer un indicateur sur l'aide sociale
}
\examples{
selectIndic(nomvariable="NbBenefAPA",denom="pop.60.99")
selectIndic(nomvariable="DepBruteAPA",denom="NbBenefAPA",options="mensuel")
selectIndic(nomvariable="NbBenefAPA",denom="pop.60.99", gpeDpt=c("Meuse","Moselle"))
}
