% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectZones.R
\name{selectZones}
\alias{selectZones}
\title{Une fonction auxiliaire qui crée les tables nécessaires pour afficher des zones (interquartiles, etc) sur un graphique}
\usage{
selectZones(nomvariable, options, nbrang = 1, tab, poidsobs)
}
\arguments{
\item{nomvariable}{le nom de la variable, parmi celles de la table ASDEPsl}

\item{options}{un vecteur d'options issu de la fonction d'origine}

\item{tab}{une table avec les valeurs de l'indicateur}

\item{poidsobs}{une variable de pondération (par exemple si on souhaite calculer des quantiles pondérés par la population des départements). Par défaut, si denom est renseigné, cette variable denom est utilisée comme poidsobs}
}
\value{

}
\description{
Cette fonction est notamment utilisée dans les fonctions graphEvolution, graphComparaison, etc.
}
